requirements.txt
twitchAPI
google-genai
python-dotenv
chromadb
ollama
streamlink
requests
beautifulsoup4
cloudscraper
pytesseract
Pillow